+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unreleased
 
+* Check if BottomSheetFragment is visible using Android Fragment API (fixes #348)
 * Fix issue where `UserCanceledException` was not being sent on bottom sheet dismissal.
 
 ## 6.1.0